Remove api for emottak-subscription
